Time: 71 ms (47.87%), Space: 98.8 MB (9.57%) - LeetHub

+++ 2448-count-number-of-bad-pairs/2448-count-number-of-bad-pairs.cpp
@@ -13,7 +13,7 @@ public:
 
         unordered_map<long long, long long> freq;
         for(int j=0; j<n; j++){
-            long long elementsSeenBeforeJ = j;
+            int elementsSeenBeforeJ = j;
             result += (elementsSeenBeforeJ - freq[modifiedNums[j]]);
 
             freq[modifiedNums[j]]++;
